First working webworker implementation
