+++ src/sources/MVT.js
@@ -122,9 +122,15 @@ export default class MVT extends Base {
 
         // Fix, not one worker per clone MVT
         this._worker = new Worker();
-        this._worker.onmessage = function (event) {
-            console.warn(event); // "This text was generated by template B"
+        this._worker.onmessage = event => {
+            const mID = event.data.mID;
+            const dataframe = event.data.dataframe;
+            Object.setPrototypeOf(dataframe, new Dataframe({}));
+            this._workerDispatch[mID](dataframe);
         };
+
+        this._workerDispatch = {};
+        this._mID = 0;
     }
 
     _clone () {
@@ -146,160 +152,16 @@ export default class MVT extends Base {
         //      add Dataframe.prototype to returned stripped dataframe on wrapper
 
         return this._tileClient.requestData(zoom, viewport, (x, y, z, templateURLs) => {
-            this._worker.postMessage({x, y, z, templateURLs});
+            return new Promise(resolve => {
+                this._worker.postMessage({ x, y, z, templateURLs, layerID: this._options.layerID, metadata: this._metadata, mID: this._mID });
+                this._workerDispatch[this._mID] = resolve;
+                this._mID++;
+            });
         },
         zoom => this._options.maxZoom === undefined
             ? this._options.viewportZoomToSourceZoom(zoom)
             : Math.min(this._options.viewportZoomToSourceZoom(zoom), this._options.maxZoom)
         );
-    }
-
-    _decodeMVTLayer (mvtLayer, metadata, mvtExtent) {
-        if (!mvtLayer.length) {
-            return { properties: [], geometries: {}, numFeatures: 0 };
-        }
-        if (!metadata.geomType) {
-            metadata.geomType = this._autoDiscoverType(mvtLayer);
-        }
-        switch (metadata.geomType) {
-            case geometryTypes.POINT:
-                return this._decode(mvtLayer, metadata, mvtExtent, new Float32Array(mvtLayer.length * 2 * 3));
-            case geometryTypes.LINE:
-                return this._decode(mvtLayer, metadata, mvtExtent, [], decodeLines);
-            case geometryTypes.POLYGON:
-                return this._decode(mvtLayer, metadata, mvtExtent, [], decodePolygons);
-            default:
-                throw new Error('MVT: invalid geometry type');
-        }
-    }
-
-    _autoDiscoverType (mvtLayer) {
-        const type = mvtLayer.feature(0).type;
-        switch (type) {
-            case mvtDecoderGeomTypes.point:
-                return geometryTypes.POINT;
-            case mvtDecoderGeomTypes.line:
-                return geometryTypes.LINE;
-            case mvtDecoderGeomTypes.polygon:
-                return geometryTypes.POLYGON;
-            default:
-                throw new Error('MVT: invalid geometry type');
-        }
-    }
-
-    _decode (mvtLayer, metadata, mvtExtent, geometries, decodeFn) {
-        let numFeatures = 0;
-        const { properties, propertyNames } = this._initializePropertyArrays(metadata, mvtLayer.length);
-        for (let i = 0; i < mvtLayer.length; i++) {
-            const f = mvtLayer.feature(i);
-            this._checkType(f, metadata.geomType);
-            const geom = f.loadGeometry();
-            if (decodeFn) {
-                const decodedPolygons = decodeFn(geom, mvtExtent);
-                geometries.push(decodedPolygons);
-            } else {
-                const x = 2 * (geom[0][0].x) / mvtExtent - 1.0;
-                const y = 2 * (1.0 - (geom[0][0].y) / mvtExtent) - 1.0;
-                // Tiles may contain points in the border;
-                // we'll avoid here duplicated points between tiles by excluding the 1-edge
-                if (x < -1 || x >= 1 || y < -1 || y >= 1) {
-                    continue;
-                }
-                geometries[6 * numFeatures + 0] = x;
-                geometries[6 * numFeatures + 1] = y;
-                geometries[6 * numFeatures + 2] = x;
-                geometries[6 * numFeatures + 3] = y;
-                geometries[6 * numFeatures + 4] = x;
-                geometries[6 * numFeatures + 5] = y;
-            }
-            if (f.properties[this._metadata.idProperty] === undefined) {
-                throw new Error(`MVT feature with undefined idProperty '${this._metadata.idProperty}'`);
-            }
-            this._decodeProperties(propertyNames, properties, f, numFeatures);
-            numFeatures++;
-        }
-
-        return { properties, geometries, numFeatures };
-    }
-
-    // Currently only mvtLayers with the same type in every feature are supported
-    _checkType (feature, expected) {
-        const type = feature.type;
-        const actual = MVT_TO_CARTO_TYPES[type];
-        if (actual !== expected) {
-            throw new Error(`MVT: mixed geometry types in the same layer. Layer has type: ${expected} but feature was ${actual}`);
-        }
-    }
-
-    _initializePropertyArrays (metadata, length) {
-        const propertyNames = this._getPropertyNamesFrom(metadata);
-        const properties = this._getPropertiesFor(propertyNames, length);
-        return { propertyNames, properties };
-    }
-
-    _getPropertyNamesFrom (metadata) {
-        const propertyNames = [];
-        for (let i = 0; i < metadata.propertyKeys.length; i++) {
-            const propertyName = metadata.propertyKeys[i];
-            if (metadata.properties[propertyName].type === 'geometry') {
-                continue;
-            }
-            propertyNames.push(...metadata.propertyNames(propertyName));
-        }
-        return propertyNames;
-    }
-
-    _getPropertiesFor (propertyNames, length) {
-        const properties = {};
-        const size = Math.ceil(length / RTT_WIDTH) * RTT_WIDTH;
-
-        const arrayBuffer = new ArrayBuffer(4 * size * propertyNames.length);
-        for (let i = 0; i < propertyNames.length; i++) {
-            const propertyName = propertyNames[i];
-            properties[propertyName] = new Float32Array(arrayBuffer, i * 4 * size, size);
-        }
-
-        return properties;
-    }
-
-    _decodeProperties (propertyNames, properties, feature, i) {
-        const length = propertyNames.length;
-        for (let j = 0; j < length; j++) {
-            const propertyName = propertyNames[j];
-            const propertyValue = feature.properties[propertyName];
-            properties[propertyName][i] = this.decodeProperty(propertyName, propertyValue);
-        }
-    }
-
-    decodeProperty (propertyName, propertyValue) {
-        if (typeof propertyValue === 'string') {
-            if (this._metadata.properties[propertyName].type !== 'category') {
-                throw new Error(`MVT decoding error. Metadata property '${propertyName}' is of type '${this._metadata.properties[propertyName].type}' but the MVT tile contained a feature property of type string: '${propertyValue}'`);
-            }
-            return this._metadata.categorizeString(propertyName, propertyValue);
-        } else if (typeof propertyValue === 'number') {
-            if (this._metadata.properties[propertyName].type !== 'number') {
-                throw new Error(`MVT decoding error. Metadata property '${propertyName}' is of type '${this._metadata.properties[propertyName].type}' but the MVT tile contained a feature property of type number: '${propertyValue}'`);
-            }
-            return propertyValue;
-        } else if (propertyValue === null || propertyValue === undefined) {
-            return Number.NaN;
-        } else {
-            throw new Error(`MVT decoding error. Feature property value of type '${typeof propertyValue}' cannot be decoded.`);
-        }
-    }
-
-    _generateDataFrame (rs, geometry, properties, size, type) {
-        return new Dataframe({
-            active: false,
-            center: rs.center,
-            geom: geometry,
-            properties: properties,
-            scale: rs.scale,
-            size: size,
-            type: type,
-            metadata: this._metadata
-        });
     }
 
     free () {
